consolidate build and build-docs
